Refactor: Improve type safety and extract context hook
