+++ src/components/ui/typography-constants.ts
@@ -1,0 +1,34 @@
+import {
+  H1,
+  H2,
+  H3,
+  H4,
+  H5,
+  H6,
+  BodyLarge,
+  Body,
+  BodySmall,
+  Caption,
+  Label,
+  Display,
+  Code,
+  Quote
+} from './typography';
+
+// Create a Typography namespace object
+export const Typography = {
+  H1,
+  H2,
+  H3,
+  H4,
+  H5,
+  H6,
+  BodyLarge,
+  Body,
+  BodySmall,
+  Caption,
+  Label,
+  Display,
+  Code,
+  Quote
+};
