+++ src/components/ui/typography-utils.tsx
@@ -16,20 +16,5 @@ export {
   Quote
 } from './typography';
 
-// Create a Typography namespace object
-export const Typography = {
-  H1,
-  H2,
-  H3,
-  H4,
-  H5,
-  H6,
-  BodyLarge,
-  Body,
-  BodySmall,
-  Caption,
-  Label,
-  Display,
-  Code,
-  Quote
-};
+// Import Typography namespace from constants file
+export { Typography } from './typography-constants';
